Revert "Alert when editing vitals of old visit"
