+++ pages/registration/index.js
@@ -17,7 +17,6 @@ import CustomModal from '@/components/CustomModal';
 import { PageTitle } from '@/components/TextComponents';
 import { REGISTRATION_FORM_FIELDS } from '@/utils/constants';
 import { VillageContext } from '@/context/VillageContext';
-import { checkRecency } from '@/utils/check';
 
 export const submitNewVisit = async patient => {
   try {
@@ -206,18 +205,41 @@ const Registration = () => {
   });
 
   const handleNewVisit = useWithLoading(async patient => {
-    const isNotRecent = await checkRecency(
-      patient.pk,
-      60,
-      true,
-      `Are you sure you want to create a new visit?
-      (Note: duplicate visits will cause unnecessary complications)
-      Press CANCEL to NOT CREATE a new visit.
-      Press OK to CREATE a new visit.`
-    );
-    if (isNotRecent) {
-      submitNewVisit(patient);
+    try {
+      const { data: patient_visits } = await axiosInstance.get(
+        `/visits?patient=${patient.pk}`
+      );
+      if (patient_visits.length > 0) {
+        const mostRecentVisit = patient_visits.reduce((latest, visit) => {
+          return new Date(visit.date) > new Date(latest.date) ? visit : latest;
+        }, patient_visits[0]);
+
+        const currentDate = new Date();
+        const mostRecentVisitDate = new Date(mostRecentVisit.date);
+        const timeDiffInMin =
+          Math.abs(currentDate.getTime() - mostRecentVisitDate.getTime()) /
+          (1000 * 60);
+
+        if (timeDiffInMin < 60) {
+          const isNotDuplicateVisit = confirm(
+            `It's only been ${Math.floor(timeDiffInMin)} min since a visit for this patient was created.
+          Are you sure you want to create a new visit?
+          (Note: duplicate visits will cause unnecessary complications)
+          Press CANCEL to NOT CREATE a new visit.
+          Press OK to CREATE a new visit.`
+          );
+
+          if (!isNotDuplicateVisit) {
+            return;
+          }
+        }
+      }
+    } catch (error) {
+      toast.error(`Error fetching patient visits: ${error.message}`);
+      console.error('Error fetching patient vists:', error);
     }
+
+    submitNewVisit(patient);
   });
 
   return (
